Adding more tests for openfermion conversion
