+++ test/operator/test_operator.py
@@ -486,71 +486,6 @@ def test_operator_on_subspace():
     assert h4.acting_on == h1.acting_on
 
 
-op_ferm = {}
-hi = nkx.hilbert.SpinOrbitalFermions(3)
-op_ferm["FermionOperator2nd_hermitian"] = (
-    nkx.operator.FermionOperator2nd(
-        hi, terms=(((0, 0), (1, 1)), ((1, 0), (0, 1))), weights=(1.0 + 1j, 1 - 1j)
-    ),
-    True,
-)
-op_ferm["FermionOperator2nd_not_hermitian"] = (
-    nkx.operator.FermionOperator2nd(
-        hi, terms=(((0, 0), (2, 1)), ((1, 0), (0, 1))), weights=(1.0 + 1j, 1 - 1j)
-    ),
-    False,
-)
-
-op_ferm["FermionOperator2nd_hermitian_3term"] = (
-    nkx.operator.FermionOperator2nd(
-        hi,
-        (((0, 0), (1, 1), (2, 1)), ((2, 0), (1, 0), (0, 1))),
-        weights=(1.0 - 1j, 1 + 1j),
-    ),
-    True,
-)
-op_ferm["FermionOperator2nd_not_hermitian_3term"] = (
-    nkx.operator.FermionOperator2nd(
-        hi,
-        (((0, 0), (1, 1), (2, 1)), ((3, 0), (1, 0), (0, 1))),
-        weights=(1.0 - 1j, 2 + 2j),
-    ),
-    False,
-)
-
-op_ferm["fermihubbard_int"] = (
-    nkx.operator.FermionOperator2nd(
-        hi,
-        terms=(
-            ((0, 1), (0, 0), (1, 1), (1, 0)),
-            ((0, 1), (0, 0), (1, 1), (1, 0)),
-            ((0, 1), (0, 0), (1, 1), (1, 0)),
-            ((0, 1), (0, 0), (1, 1), (1, 0)),
-        ),
-        weights=(1.0, 1.0, 1.0, 1.0),
-    ),
-    True,
-)
-
-op_ferm["ordering"] = (
-    nkx.operator.FermionOperator2nd(
-        hi,
-        terms=(((0, 1), (0, 0), (1, 1), (1, 0)), ((1, 1), (1, 0), (0, 1), (0, 0))),
-        weights=(1.0 - 1j, 1 + 1j),
-    ),
-    True,
-)
-
-
-@pytest.mark.parametrize(
-    "op_ferm, is_hermitian",
-    [pytest.param(op, is_herm, id=name) for name, (op, is_herm) in op_ferm.items()],
-)
-def test_is_hermitian_fermion2nd(op_ferm, is_hermitian):
-    print("OPERATOR", op_ferm._operators)
-    assert op_ferm.is_hermitian == is_hermitian
-
-
 def test_openfermion_conversion():
     # skip test if openfermion not installed
     pytest.importorskip("openfermion")
@@ -616,274 +551,3 @@ def test_openfermion_conversion():
     assert fo2.hilbert.size == 3 * 2
     # will fail of we go outside of the allowed states with openfermion operators
     fo2.to_dense()
-
-
-def test_fermion_operator_with_strings():
-    hi = nkx.hilbert.SpinOrbitalFermions(3)
-    terms = (((0, 1), (2, 0)),)
-    op1 = nkx.operator.FermionOperator2nd(hi, terms)
-    op2 = nkx.operator.FermionOperator2nd(hi, ("0^ 2",))
-    assert np.allclose(op1.to_dense(), op2.to_dense())
-
-    terms = (((0, 1), (1, 0)), ((2, 1), (1, 0)))
-    weights = (0.5 - 0.5j, 0.5 + 0.5j)
-    op1 = nkx.operator.FermionOperator2nd(hi, terms, weights)
-    op2 = nkx.operator.FermionOperator2nd(hi, ("0^ 1", "2^ 1"), weights)
-    assert np.allclose(op1.to_dense(), op2.to_dense())
-
-    terms = (((0, 1), (1, 0), (2, 1)), ((2, 1), (1, 0), (0, 1)))
-    weights = (0.5 - 0.5j, 0.5 + 0.5j)
-    op1 = nkx.operator.FermionOperator2nd(hi, terms, weights)
-    op2 = nkx.operator.FermionOperator2nd(hi, ("0^ 1 2^", "2^ 1 0^"), weights)
-    assert np.allclose(op1.to_dense(), op2.to_dense())
-
-
-def compare_openfermion_fermions():
-    # skip test if openfermion not installed
-    pytest.importorskip("openfermion")
-    from openfermion import FermionOperator, get_sparse_operator
-
-    # openfermion
-    of = FermionOperator("0^ 1", 1.0) + FermionOperator("1^ 0", 1.0)
-    of_dense = get_sparse_operator(of).todense()
-    # from_openfermion
-    fo = nkx.operator.FermionOperator2nd.from_openfermion(of)
-    fo_dense = fo.to_dense()
-    # FermionOperator2nd
-    hi = nkx.hilbert.SpinOrbitalFermions(2)  # two sites
-    fermop = nkx.operator.FermionOperator2nd(
-        hi, terms=(((0, 1), (1, 0)), ((1, 1), (0, 0))), weights=(1.0, 1.0)
-    )
-    fermop_dense = fermop.to_dense()
-    # compare openfermion vs from_openfermion
-    assert np.array_equal(of_dense, fo_dense)
-    # compare openfermion vs FermionOperator2nd
-    assert np.array_equal(of_dense, fermop_dense)
-    # compare from_openfermion vs FermionOperator 2nd
-    assert np.array_equal(fo_dense, fermop_dense)
-
-
-def test_add_fermions():
-    hi = nkx.hilbert.SpinOrbitalFermions(5)
-    op1 = nkx.operator.FermionOperator2nd(hi, terms=("1^ 2"), weights=(1,), constant=2)
-    op2 = nkx.operator.FermionOperator2nd(
-        hi, terms=("3^ 4"), weights=(1.3,), constant=5.7
-    )
-    op3 = nkx.operator.FermionOperator2nd(
-        hi, terms=("3^ 4", "1^ 2"), weights=(1.3, 1), constant=7.7
-    )
-    op4 = op3 * 2
-    op5 = nkx.operator.FermionOperator2nd(
-        hi, terms=("3^ 4", "1^ 2"), weights=(2 * 1.3, 2 * 1), constant=2 * 7.7
-    )
-    assert np.allclose((op1 + op2).to_dense(), op3.to_dense())
-    assert np.allclose(op4.to_dense(), op5.to_dense())
-
-
-def test_create_annihil_number():
-    hi = nkx.hilbert.SpinOrbitalFermions(5)
-    op1 = nkx.operator.FermionOperator2nd(hi, terms=("0^ 0", "1^ 2"), weights=(0.3, 2))
-
-    from netket.experimental.operator.fermion import destroy, create, number
-
-    def c(site):
-        return destroy(hi, site)
-
-    def cdag(site):
-        return create(hi, site)
-
-    def cn(site):
-        return number(hi, site)
-
-    op2 = 0.3 * cn(0) + 2 * cdag(1) * c(2)
-    assert np.allclose(op1.to_dense(), op2.to_dense())
-    op3 = nkx.operator.FermionOperator2nd(
-        hi, terms=("0^ 1", "1^ 2"), weights=(1 + 1j, 2 - 2j), constant=2
-    )
-    op4 = (1 + 1j) * cdag(0) * c(1) + (2 - 2j) * cdag(1) * c(2) + 2
-    assert np.allclose(op3.to_dense(), op4.to_dense())
-
-
-def test_operations_fermions():
-    hi = nkx.hilbert.SpinOrbitalFermions(5)
-    op1 = nkx.operator.FermionOperator2nd(
-        hi, terms=("1^ 2",), weights=(1,), constant=2, dtype=complex
-    )
-    op2 = nkx.operator.FermionOperator2nd(
-        hi, terms=("3^ 4"), weights=(1.3,), constant=5.7
-    )
-    op2copy = op2.copy()
-    assert op2copy.hilbert == op2.hilbert
-    assert np.allclose(list(op2._operators.keys()), list(op2copy._operators.keys()))
-    assert np.allclose(list(op2._operators.values()), list(op2copy._operators.values()))
-    assert op2.is_hermitian == op2copy.is_hermitian
-    assert np.allclose(op2.to_dense(), op2copy.to_dense())
-
-    op3 = nkx.operator.FermionOperator2nd(
-        hi, terms=("3^ 4", "1^ 2"), weights=(1.3, 1), constant=7.7
-    )
-    op12 = op1.copy()
-    op12 += op2
-    assert np.allclose((op1 + op2).to_dense(), op3.to_dense())
-    assert np.allclose(op12.to_dense(), op3.to_dense())
-
-    op4 = op3 * 2
-    op5 = nkx.operator.FermionOperator2nd(
-        hi, terms=("3^ 4", "1^ 2"), weights=(2 * 1.3, 2 * 1), constant=2 * 7.7
-    )
-    op4b = op3.copy()
-    op4b *= 2
-    assert np.allclose(op4.to_dense(), op5.to_dense())
-    assert np.allclose(op4b.to_dense(), op5.to_dense())
-
-    op6 = nkx.operator.FermionOperator2nd(
-        hi, terms=("1^ 2", "0^ 1"), weights=(1j, -1.0j), constant=7.7
-    )
-    op7 = nkx.operator.FermionOperator2nd(
-        hi, terms=("1^ 2", "0^ 1"), weights=(1, 1), constant=7.7
-    )
-    op8 = nkx.operator.FermionOperator2nd(
-        hi, terms=("1^ 2", "0^ 1"), weights=(1.0 + 1j, 1 - 1j), constant=2 * 7.7
-    )
-    op67 = op6.copy()
-    op67 += op7
-    assert np.allclose((op6 + op7).to_dense(), op8.to_dense())
-    assert np.allclose(op67.to_dense(), op8.to_dense())
-
-    op8 = nkx.operator.FermionOperator2nd(
-        hi, terms=("0^ 1", "2^ 3"), weights=(1 + 1j, 2 - 0.5j), constant=1.0 + 3j
-    )
-    op8_trueconj = nkx.operator.FermionOperator2nd(
-        hi, terms=("1^ 0", "3^ 2"), weights=(1 - 1j, 2 + 0.5j), constant=1.0 - 3j
-    )
-    assert np.allclose(op8.conjugate().to_dense(), op8_trueconj.to_dense())
-
-
-def test_fermion_op_matmul():
-    hi = nkx.hilbert.SpinOrbitalFermions(3)
-    op1 = nkx.operator.FermionOperator2nd(hi, terms=("0^ 0", "1^ 2"), weights=(0.3, 2))
-
-    # multiply with a real constant
-    op_real = nkx.operator.FermionOperator2nd(hi, [], [], constant=2.0)
-    assert np.allclose((op1 @ op_real).to_dense(), (op1 * 2).to_dense())
-    assert np.allclose((op1 * op_real).to_dense(), (op1 * 2).to_dense())
-
-    # multiply with a real+complex constant
-    op_complex = nkx.operator.FermionOperator2nd(hi, [], [], constant=2.0 + 2j)
-    assert np.allclose((op1 @ op_complex).to_dense(), (op1 * (2 + 2j)).to_dense())
-    assert np.allclose((op1 * op_complex).to_dense(), (op1 * (2 + 2j)).to_dense())
-
-    # multiply with another operator
-    op2 = nkx.operator.FermionOperator2nd(
-        hi, terms=("1^ 1", "0^ 2"), weights=(1 + 1j, 0.5)
-    )
-    op2b = nkx.operator.FermionOperator2nd(
-        hi,
-        terms=("0^ 0 1^ 1", "0^ 0 0^ 2", "1^ 2 1^ 1", "1^ 2 0^ 2"),
-        weights=(0.3 * (1 + 1j), 0.3 * 0.5, 2 * (1 + 1j), 2 * 0.5),
-    )
-    assert np.allclose(
-        (op1 @ op2).to_dense(),
-        op2b.to_dense(),
-    )
-    assert np.allclose(
-        (op1 * op2).to_dense(),
-        op2b.to_dense(),
-    )
-
-    # multiply with another operator + constant
-    op3 = nkx.operator.FermionOperator2nd(
-        hi, terms=("1^ 1",), weights=(1 + 1j,), constant=5
-    )
-    op3b = nkx.operator.FermionOperator2nd(
-        hi,
-        terms=("0^ 0 1^ 1", "0^ 0", "1^ 2 1^ 1", "1^ 2"),
-        weights=(0.3 * (1 + 1j), 5 * 0.3, 2 * (1 + 1j), 10),
-        constant=0,
-    )
-    assert np.allclose(
-        (op1 @ op3).to_dense(),
-        op3b.to_dense(),
-    )
-    assert np.allclose(
-        (op1 * op3).to_dense(),
-        op3b.to_dense(),
-    )
-
-
-def test_fermion_add_sub_mul():
-    # check addition
-    hi = nkx.hilbert.SpinOrbitalFermions(3)
-    op1 = nkx.operator.FermionOperator2nd(
-        hi, terms=("0^ 0", "1^ 2"), weights=(0.3, 2), constant=2
-    )
-    assert np.allclose((op1 + op1).to_dense(), 2 * op1.to_dense())
-
-    op2 = nkx.operator.FermionOperator2nd(
-        hi, terms=("0^ 0", "0^ 1"), weights=(0.5, 4j), constant=1
-    )
-    op2b = nkx.operator.FermionOperator2nd(
-        hi, terms=("0^ 0", "1^ 2", "0^ 1"), weights=(0.3 + 0.5, 2, 4j), constant=3
-    )
-    assert np.allclose((op1 + op2).to_dense(), op2b.to_dense())
-    op2c = op2.copy()
-    op2c += op1
-    assert np.allclose(op2c.to_dense(), op2b.to_dense())
-
-    # check substraction
-    op2d = nkx.operator.FermionOperator2nd(
-        hi, terms=("0^ 0", "1^ 2", "0^ 1"), weights=(0.3 - 0.5, 2, -4j), constant=1
-    )
-    assert np.allclose((op1 - op2).to_dense(), op2d.to_dense())
-    op1c = op1.copy()
-    op1c -= op2
-    assert np.allclose(op1c.to_dense(), op2d.to_dense())
-
-    # check multiplication with scalar
-    op1f = nkx.operator.FermionOperator2nd(
-        hi,
-        terms=("0^ 0", "1^ 2"),
-        weights=(
-            3,
-            20,
-        ),
-        constant=20,
-    )
-    op1c = op1.copy()
-    op1c *= 10
-    assert np.allclose((op1 * 10).to_dense(), op1f.to_dense())
-    assert np.allclose(op1c.to_dense(), op1f.to_dense())
-
-
-@pytest.mark.parametrize("dtype1", [np.float32, np.float64])
-@pytest.mark.parametrize("dtype2", [np.float32, np.float64])
-def test_dtype_promotion(dtype1, dtype2):
-    hi = nkx.hilbert.SpinOrbitalFermions(3)
-    op1 = nkx.operator.FermionOperator2nd(
-        hi, terms=("0^ 0", "1^ 2"), weights=(0.3, 2), constant=2, dtype=dtype1
-    )
-    op2 = nkx.operator.FermionOperator2nd(
-        hi, terms=("0^ 1"), weights=(0.1,), constant=2, dtype=dtype2
-    )
-
-    assert op1.dtype == dtype1
-    assert op2.dtype == dtype2
-    assert op1.to_dense().dtype == dtype1
-    assert op2.to_dense().dtype == dtype2
-
-    assert (-op1).dtype == dtype1
-    assert (-op2).dtype == dtype2
-
-    assert (op1 + op2).dtype == np.promote_types(op1.dtype, op2.dtype)
-    assert (op1 - op2).dtype == np.promote_types(op1.dtype, op2.dtype)
-    assert (op1 @ op2).dtype == np.promote_types(op1.dtype, op2.dtype)
-
-    a = np.array(0.5, dtype=dtype1)
-    assert (op2 + a + op2).dtype == np.promote_types(a.dtype, op2.dtype)
-    assert (op2 - a).dtype == np.promote_types(a.dtype, op2.dtype)
-    assert (op2 * a).dtype == np.promote_types(a.dtype, op2.dtype)
-
-    a = np.array(0.5, dtype=dtype2)
-    assert (op1 + a).dtype == np.promote_types(op1.dtype, a.dtype)
-    assert (op1 - a).dtype == np.promote_types(op1.dtype, a.dtype)
-    assert (op1 * a).dtype == np.promote_types(op1.dtype, a.dtype)
